[TECH-293] Move steps collection to separate file
branch: feature/TECH-293-custom-stages
