+++ src/mpyl/steps/steps.py
@@ -12,19 +12,8 @@ from unittest import TestSuite
 from ruamel.yaml import YAML  # type: ignore
 
 from . import Step
-from .build.dockerbuild import BuildDocker
-from .build.echo import BuildEcho
-from .build.sbt import BuildSbt
-from .deploy.echo import DeployEcho
-from .deploy.ephemeral_docker_deploy import EphemeralDockerDeploy
-from .deploy.kubernetes import DeployKubernetes
-from .deploy.kubernetes_job import DeployKubernetesJob
-from .deploy.kubernetes_spark_job import DeployKubernetesSparkJob
+from .collection import StepsCollection
 from .models import Output, Input, RunProperties, ArtifactType, Artifact
-from .postdeploy.cypress_test import CypressTest
-from .test.dockertest import TestDocker
-from .test.echo import TestEcho
-from .test.sbt import TestSbt
 from ..project import Project
 from ..project import Stage
 from ..utilities.junit import to_test_suites
@@ -64,31 +53,7 @@ def collect_test_results(test_artifacts: list[Artifact]) -> list[TestSuite]:
     return list(itertools.chain(*suites))
 
 
-class StepsCollection:
-    _step_executors: set[Step]
 
-    def __init__(self, logger: Logger) -> None:
-        self._step_executors = {
-            BuildEcho(logger),
-            BuildSbt(logger),
-            BuildDocker(logger),
-            TestEcho(logger),
-            TestSbt(logger),
-            TestDocker(logger),
-            DeployEcho(logger),
-            DeployKubernetes(logger),
-            DeployKubernetesJob(logger),
-            DeployKubernetesSparkJob(logger),
-            EphemeralDockerDeploy(logger),
-            CypressTest(logger)
-        }
-
-    def add_executor(self, step: Step):
-        self._step_executors.add(step)
-
-    def get_executor(self, stage: Stage, step_name: str) -> Optional[Step]:
-        executors = filter(lambda e: step_name == e.meta.name and e.meta.stage == stage.name, self._step_executors)
-        return next(executors, None)
 
 
 class Steps:
@@ -154,18 +119,18 @@ class Steps:
         return None
 
     def _execute_stage(self, stage: Stage, project: Project, dry_run: bool = False) -> Output:
-        stage_name = project.stages.for_stage(stage.name)
-        if stage_name is None:
+        step_name = project.stages.for_stage(stage.name)
+        if step_name is None:
             return Output(success=False, message=f"Stage '{stage.name}' not defined on project '{project.name}'")
 
         invalid_maintainers = self._validate_project_against_config(project)
         if invalid_maintainers:
             return invalid_maintainers
 
-        executor: Optional[Step] = self._steps_collection.get_executor(stage, stage_name)
+        executor: Optional[Step] = self._steps_collection.get_executor(stage.name, step_name)
         if executor is None:
-            self._logger.warning(f"No executor found for {stage_name} in stage {stage}")
-            return Output(success=False, message=f"Executor '{stage_name}' for '{stage.name}' not known or registered")
+            self._logger.warning(f"No executor found for {step_name} in stage {stage}")
+            return Output(success=False, message=f"Executor '{step_name}' for '{stage.name}' not known or registered")
 
         try:
             self._logger.info(f'Executing {stage} for {project.name}')
